pyproject.toml
[tool.poetry]
name = "x-ui-telegram"
version = "0.1.0"
description = ""
authors = ["Ehsan <e.me@gmail.com>"]
readme = "README.md"
packages = [{include = "x_ui_telegram"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.2"
python-telegram-bot = {extras = ["job-queue"], version = "^20.1"}
persiantools = "^3.0.1"
pytz = "^2022.7.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
